Remove all global opcodes
The implementation of shecc uses several 'global' opcodes
to handle the operations such as load/store global variables
and then output proper instructions for them.

However, this commit removes all 'global' opcodes and uses
phase2 IR to transfer a flag, which determines whether the
operand is global, and generate the instructions eventually.

Thus, shecc doesn't need to handle the variant of the similar
opcodes.